Refactor project and upload file

+++ src/T2M_GPT/utils/quaternion.py
@@ -5,12 +5,12 @@
 # LICENSE file in the root directory of this source tree.
 #
 
-import torch
 import numpy as np
+import torch
 
 _EPS4 = np.finfo(float).eps * 4.0
 
-_FLOAT_EPS = np.finfo(np.float).eps
+_FLOAT_EPS = np.finfo(float).eps
 
 
 # PyTorch-backed implementations
